Add :someone (properly)
